requirements/base.txt
python-dotenv==1.0.0
Django==4.1.3
python-dotenv==1.0.0
requests==2.22.0
docker==4.1.0